Build: Bump gradle-baseline-java to 6.90.0

+++ parquet/src/main/java/org/apache/iceberg/parquet/Parquet.java
@@ -477,7 +477,7 @@ public class Parquet {
                 .setWriteSupport(getWriteSupport(type))
                 .withCompressionCodec(codec)
                 .withWriteMode(writeMode)
-                .withRowGroupSize(rowGroupSize)
+                .withRowGroupSize((long) rowGroupSize)
                 .withPageSize(pageSize)
                 .withPageRowCountLimit(pageRowLimit)
                 .withDictionaryEncoding(dictionaryEnabled)
